requirements.txt
influxdb-client[ciso]>=1.40.0
questdb>=1.2.0
clickhouse-connect>=0.7.0
numpy>=1.24.0
pandas>=2.0.0
paho-mqtt<2.0.0
python-dotenv>=1.0.0
